chore: rename lambda to IR (#2030)
