+++ src/core/blueprint/from_config.rs
@@ -7,8 +7,8 @@ use super::{Server, TypeLike};
 use crate::core::blueprint::compress::compress;
 use crate::core::blueprint::*;
 use crate::core::config::{Arg, Batch, Config, ConfigModule, Field};
+use crate::core::ir::{IO, IR};
 use crate::core::json::JsonSchema;
-use crate::core::lambda::{Expression, IO};
 use crate::core::try_fold::TryFold;
 use crate::core::valid::{Valid, ValidationError, Validator};
 
@@ -56,9 +56,7 @@ pub fn apply_batching(mut blueprint: Blueprint) -> Blueprint {
     for def in blueprint.definitions.iter() {
         if let Definition::Object(object_type_definition) = def {
             for field in object_type_definition.fields.iter() {
-                if let Some(Expression::IO(IO::Http { group_by: Some(_), .. })) =
-                    field.resolver.clone()
-                {
+                if let Some(IR::IO(IO::Http { group_by: Some(_), .. })) = field.resolver.clone() {
                     blueprint.upstream.batch = blueprint.upstream.batch.or(Some(Batch::default()));
                     return blueprint;
                 }
