+++ src/core/blueprint/operators/modify.rs
@@ -1,7 +1,7 @@
 use crate::core::blueprint::*;
 use crate::core::config;
 use crate::core::config::Field;
-use crate::core::lambda::{Context, Expression};
+use crate::core::ir::{Context, IR};
 use crate::core::try_fold::TryFold;
 use crate::core::valid::Valid;
 
@@ -22,9 +22,11 @@ pub fn update_modify<'a>(
                             }
                         }
                     }
-                    b_field.resolver = Some(b_field.resolver.unwrap_or(Expression::Context(
-                        Context::Path(vec![b_field.name.clone()]),
-                    )));
+                    b_field.resolver = Some(
+                        b_field
+                            .resolver
+                            .unwrap_or(IR::Context(Context::Path(vec![b_field.name.clone()]))),
+                    );
                     b_field = b_field.name(new_name.clone());
                 }
             }
