+++ src/core/blueprint/operators/protected.rs
@@ -1,6 +1,6 @@
 use crate::core::blueprint::FieldDefinition;
 use crate::core::config::{self, ConfigModule, Field};
-use crate::core::lambda::{Context, Expression};
+use crate::core::ir::{Context, IR};
 use crate::core::try_fold::TryFold;
 use crate::core::valid::Valid;
 
@@ -27,9 +27,10 @@ pub fn update_protected<'a>(
                     );
                 }
 
-                b_field.resolver = Some(Expression::Protect(Box::new(b_field.resolver.unwrap_or(
-                    Expression::Context(Context::Path(vec![b_field.name.clone()])),
-                ))));
+                b_field.resolver =
+                    Some(IR::Protect(Box::new(b_field.resolver.unwrap_or(
+                        IR::Context(Context::Path(vec![b_field.name.clone()])),
+                    ))));
             }
 
             Valid::succeed(b_field)
