+++ src/core/mod.rs
@@ -18,8 +18,8 @@ pub mod grpc;
 pub mod has_headers;
 pub mod helpers;
 pub mod http;
+pub mod ir;
 pub mod json;
-pub mod lambda;
 pub mod merge_right;
 pub mod mustache;
 pub mod path;
